fix: Add installation output and increase sentinel lifetime
Currently, 'helm install' does not only not reveal much information about the progress but also does not proceed anymore as soon as the sentinel has completed. Thus, in case the sentinel times out without connaisseur being ready, the last 4 minutes waiting for the helm timeout are wasted. This adds a bit output during installation to imprive guidance of the user and increases the sentinel's lifetime to 5 minutes. As the bootstrap pod will be killed as soon as notary as well as webhook are both ready, there is no need to keep it small. Thus, we went for 5 minutes which is the same  as the helm timeout.