fix: fixed delegation validation behavior
Prviously, when delegations were present in the trust data, only the delegation JSONs were considered for extracting the digest, which might result in multiple digests for the same tag and thus in an error. Now, if delegations are present, the releases JSON will be looked at exclusively, should none of the delegations be explicitly required.

fixes #61